refactor(load-test): split request building/parsing into a separate module
